Fix most pytest warnings

+++ qasync/__init__.py
@@ -244,7 +244,7 @@ class _SimpleTimer(QtCore.QObject):
         self._stopped = False
 
     def add_callback(self, handle, delay=0):
-        timerid = self.startTimer(delay * 1000)
+        timerid = self.startTimer(int(delay * 1000))
         self._logger.debug("Registering timer id {0}".format(timerid))
         assert timerid not in self.__callbacks
         self.__callbacks[timerid] = handle
